pyproject.toml
[tool.poetry]
name = "ad-campaign-manager-project"
version = "0.1.0"
description = ""
authors = ["Regina0101 <ritodo1010@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1"
djangorestframework = "^3.15.2"
psycopg2-binary = "^2.9.9"
django-crispy-forms = "^2.3"
django-allauth = "^64.2.0"
crispy-bootstrap5 = "^2024.2"
google-auth-oauthlib = "^1.2.1"
google-auth-httplib2 = "^0.2.0"
pyjwt = "^2.9.0"
cryptography = "^43.0.1"
transformers = "^4.44.2"
torch = "^2.4.1"
generativeai = "^0.0.1"
google-generativeai = "^0.8.0"
django-anymail = "^12.0"
sib-api-v3-sdk = "^7.6.0"
diffusers = "^0.30.2"
accelerate = "^0.34.2"
torchvision = "^0.19.1"
torchaudio = "^2.4.1"
requests = "^2.32.3"
pillow = "^10.4.0"
python-dotenv = "^1.0.1"
matplotlib = "^3.9.2"
pytrends = "^4.9.2"
pytest = "^8.3.3"
pytest-django = "^4.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
